+++ crates/cli/src/commands/plugin/info.rs
@@ -18,7 +18,7 @@ struct InfoPluginResult {
     bins: Vec<ExecutableLocation>,
     config: ProtoToolConfig,
     exe_file: PathBuf,
-    exes_dir: Option<PathBuf>,
+    exes_dirs: Vec<PathBuf>,
     globals_dirs: Vec<PathBuf>,
     globals_prefix: Option<String>,
     id: Id,
@@ -61,7 +61,7 @@ pub async fn info(session: ProtoSession, args: InfoPluginArgs) -> AppResult {
             bins,
             config: tool.config.clone(),
             exe_file: tool.locate_exe_file().await?,
-            exes_dir: tool.locate_exes_dir().await?,
+            exes_dirs: tool.locate_exes_dirs().await?,
             globals_dirs: tool.locate_globals_dirs().await?,
             globals_prefix: tool.locate_globals_prefix().await?,
             inventory_dir: tool.get_inventory_dir(),
@@ -165,7 +165,7 @@ pub async fn info(session: ProtoSession, args: InfoPluginArgs) -> AppResult {
     // INVENTORY
 
     let exe_file = tool.locate_exe_file().await?;
-    let exes_dir = tool.locate_exes_dir().await?;
+    let exes_dirs = tool.locate_exes_dirs().await?;
     let globals_dir = tool.locate_globals_dir().await?;
     let globals_prefix = tool.locate_globals_prefix().await?;
 
@@ -212,19 +212,23 @@ pub async fn info(session: ProtoSession, args: InfoPluginArgs) -> AppResult {
                         )
                     }.into_any()
                 )
-                #(exes_dir.map(|dir| {
-                    element! {
-                        Entry(
-                            name: "Executables directory",
-                            value: element! {
-                                StyledText(
-                                    content: dir.to_string_lossy(),
-                                    style: Style::Path
-                                )
-                            }.into_any()
-                        )
+                Entry(
+                    name: "Executables directories",
+                    no_children: exes_dirs.is_empty()
+                ) {
+                    List {
+                        #(exes_dirs.into_iter().map(|dir| {
+                            element! {
+                                ListItem {
+                                    StyledText(
+                                        content: dir.to_string_lossy(),
+                                        style: Style::Path
+                                    )
+                                }
+                            }
+                        }))
                     }
-                }))
+                }
                 #(globals_prefix.map(|prefix| {
                     element! {
                         Entry(